README.md
Its a game based on the user's input, played between two players. And to win this game, both of them "player 1" and "player 2" are trying to set their own marks ("0" and "X") respectively with each other in  (side manner) or from (top to bottom manner) and also in the (cross manner) to make 3 "0"s or 3 "X"s are adjacent to each other. So who will do this first, then we will get the winner of this game.